SetupHelp.cs: check directory before removing
Avoid throwing errors when directory does not exist.  Check before
delete.

+++ Quandl.Excel.Console/SetupHelp.cs
@@ -150,7 +150,10 @@ namespace Quandl.Excel.Console
         private static void CleanLogs()
         {
             var documentsPath = Path.Combine(Environment.ExpandEnvironmentVariables("%userprofile%"), "Documents", "Nasdaq");
-            Directory.Delete(documentsPath, true);
+            if (Directory.Exists(documentsPath))
+            {
+                Directory.Delete(documentsPath, true);
+            }
         }
